requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-docx
httpx
python-dotenv
python-multipart
